main 분리, header 색변경
